refactor: ContentLayer를 통해 Portal과 관심사 분리
